Ui fixes and code clean-up
Cleaned up the code in ManualActivity.
Cleaned up the UI by changing the name of the app title and removing old pages.
Added some comments here and there to explain the code a bit.
Moved Client and MqttClient to a utility package.
Fixed some minor issues in the .xml files